requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
stripe==7.8.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
